% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxy2HTML.R
\name{roxy2HTML}
\alias{roxy2HTML}
\title{roxy2HTML}
\usage{
roxy2HTML(input, output, env = NULL, ...)
}
\arguments{
\item{input, }{character, path to R file with roxygen2 tags}

\item{output, }{character, path to output file to save HTML files}

\item{env, }{environment, environment to run parse_file.  Defaults to \code{NULL}}

\item{..., }{additional parameters to be passed to Rd2HTML}
}
\value{
This function doesn't return anything
}
\description{
Convert roxygen2 tags directly to html
}
\details{
Takes, as input, an R script with roxygen2 tags and outputs the
HTML file to an output folder
}
\examples{
input_path <- system.file("examples/script.R", package = "directRoxy")
roxy2HTML(input_path, tempdir())

}
