% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxy2Rd.R
\name{roxy2Rd}
\alias{roxy2Rd}
\title{roxy2Rd}
\usage{
roxy2Rd(input, output, env = NULL)
}
\arguments{
\item{input, }{character, path to R file with roxygen2 tags}

\item{output, }{character, path to output file to save Rd files}

\item{env, }{environment, environment to run parse_file.  Defaults to \code{NULL}}
}
\value{
This function doesn't return anything
}
\description{
Convert roxygen2 tags directly to Rd
}
\details{
Takes, as input, an R script with roxygen2 tags and outputs the
Rd file to an output folder
}
\examples{
# roxy2Rd("/path/to/script.R", "/path/output/")

}
